requirements.txt
django
django-widget-tweaks
django-tempus-dominus
